+++ magda-web-client/src/Components/Search/SearchBox.js
@@ -9,12 +9,11 @@ import React, { Component } from "react";
 import { fetchRegionMapping } from "../../actions/regionMappingActions";
 import Form from "muicss/lib/react/form";
 import Input from "muicss/lib/react/input";
-import search from "../../assets/search-white.svg";
+import searchGrey from "../../assets/search-grey.svg";
+import searchDark from "../../assets/search-dark.svg";
 // eslint-disable-next-line
 import PropTypes from "prop-types";
 import queryString from "query-string";
-// import Particles from '../../UI/Particles';
-import SearchSuggestionBox from "./SearchSuggestionBox";
 
 class SearchBox extends Component {
     constructor(props) {
@@ -135,62 +134,35 @@ class SearchBox extends Component {
     render() {
         return (
             <Form className="searchBox">
-                <table width="100%">
-                    <tbody>
-                        <tr>
-                            <td>
-                                <label htmlFor="search">
-                                    <span className="sr-only">
-                                        {"search " + config.appName}
-                                    </span>
-                                    <div style={{ position: "relative" }}>
-                                        <Input
-                                            type="text"
-                                            name="search"
-                                            id="search"
-                                            placeholder="search for open data"
-                                            value={this.getSearchBoxValue()}
-                                            onChange={this.onSearchTextChange}
-                                            onKeyPress={
-                                                this
-                                                    .handleSearchFieldEnterKeyPress
-                                            }
-                                            onFocus={() =>
-                                                this.setState({ isFocus: true })
-                                            }
-                                            onBlur={() =>
-                                                this.setState({
-                                                    isFocus: false
-                                                })
-                                            }
-                                            autoComplete="off"
-                                        />
-                                        <SearchSuggestionBox
-                                            searchText={this.getSearchBoxValue()}
-                                            isSearchInputFocus={
-                                                this.state.isFocus
-                                            }
-                                        />
-                                    </div>
-                                </label>
-                            </td>
-                            <td className="search-btn-box">
-                                <button
-                                    onClick={this.onClickSearch}
-                                    className="search-btn"
-                                    type="button"
-                                >
-                                    <img
-                                        src={search}
-                                        alt="search button"
-                                        title="Search"
-                                    />
-                                    <span className="sr-only">Search</span>
-                                </button>
-                            </td>
-                        </tr>
-                    </tbody>
-                </table>
+                <label htmlFor="search">
+                    <span className="sr-only">
+                        {"Search " + config.appName}
+                    </span>
+                    <Input
+                        type="text"
+                        name="search"
+                        id="search"
+                        placeholder="Search for open data"
+                        value={this.getSearchBoxValue()}
+                        onChange={this.onSearchTextChange}
+                        onKeyPress={this.handleSearchFieldEnterKeyPress}
+                        autoComplete="off"
+                    />
+                    <span className="search-input__highlight">
+                        {this.getSearchBoxValue()}
+                    </span>
+                </label>
+                <button
+                    onClick={this.onClickSearch}
+                    className="search-btn"
+                    type="button"
+                >
+                    <img
+                        src={this.props.isMobile ? searchDark : searchGrey}
+                        alt="search button"
+                    />
+                    <span className="sr-only">submit search</span>
+                </button>
             </Form>
         );
     }