+++ CHANGES.md
@@ -25,6 +25,9 @@
 * Hid download button on dataset page when download url is not available
 * Updated footer links and layout
 * Fixed an issue that prevents csv-geo-au data source to be opened in national map
+* responsive background image for homepage
+* Hidden top `go back to old site` for mobile views
+* New Homepage Design
 * Open in National Map button now send config via postMessage (except IE <=11)
 
 ## 0.0.36